+++ gtas-parent/gtas-commons/src/main/java/gov/gtas/summary/MessageTravelInformation.java
@@ -1,4 +1,4 @@
-package gov.gtas.job.scheduler.summary;
+package gov.gtas.summary;
 
 import gov.gtas.model.BookingDetail;
 import gov.gtas.model.Flight;
