+++ gtas-parent/gtas-job-scheduler-war/src/main/java/gov/gtas/job/scheduler/LoaderQueueThreadManager.java
@@ -9,6 +9,7 @@ import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import gov.gtas.job.config.JobSchedulerConfig;
+import gov.gtas.summary.EventIdentifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
